fix: move the cloud specific functions from cmd into a cloud package
