+++ pkg/cloud/aks/helper.go
